+++ examples/python/dna/main.py
@@ -18,31 +18,8 @@ DESCRIPTION = """
 This is a minimal example that logs synthetic 3D data in the shape of a double helix. The underlying data is generated
 using numpy and visualized using Rerun.
 
-## How it was made
 The full source code for this example is available
-[on GitHub](https://github.com/rerun-io/rerun/blob/latest/examples/python/dna/main.py).
-
-### Colored 3D points
-The colored 3D points were added to the scene by logging the
-[rr.Points3D archetype](https://www.rerun.io/docs/reference/types/archetypes/points3d) to the
-[helix/structure/left](recording://helix/structure/left) and [helix/structure/right](recording://helix/structure/right)
-entities.
-
-### 3D line strips
-The 3D line strips connecting the 3D point pairs are logged as an
-[rr.LineStrips3D archetype](https://www.rerun.io/docs/reference/types/archetypes/line_strips3d) to the
-[helix/structure/scaffolding entity](recording://helix/structure/scaffolding).
-
-### Rotation
-The whole structure is rotated over time by logging a
-[rr.Transform3D archetype](https://www.rerun.io/docs/reference/types/archetypes/transform3d) to the
-[helix/structure entity](recording://helix/structure:Transform3D) that changes over time. This transform determines the rotation of
-the [structure entity](recording://helix/structure) relative to the [helix](recording://helix) entity. Since all other
-entities are children of [helix/structure](recording://helix/structure) they will also rotate based on this transform.
-
-You can visualize this rotation by selecting the two entities on the left-hand side and activating `Show transform` in
-the Blueprint settings on the right-hand side. You will see one static frame (i.e., the frame of
-[helix](recording://helix)) and the rotating frame (i.e., the frame of [structure](recording://helix/structure)).
+[on GitHub](https://github.com/rerun-io/rerun/blob/latest/examples/python/dna).
 """.strip()