+++ rerun_py/rerun_sdk/rerun/any_value.py
@@ -102,7 +102,8 @@ class AnyValues(AsComponents):
     ```python
     rr.log(
         "any_values", rr.AnyValues(
-            foo=[1.2, 3.4, 5.6], bar="hello world",
+            confidence=[1.2, 3.4, 5.6],
+            description="Bla bla bla…",
         ),
     )
     ```